+++ bin/config.json
@@ -1,0 +1,15 @@
+{
+    "pulseha": {
+        "tls": false,
+        "hcs_interval": 1000,
+        "fos_interval": 5000,
+        "fo_limit": 10000
+    },
+    "floating_ip_groups": {},
+    "nodes": {},
+    "logging": {
+        "level": "info",
+        "to_logfile": false,
+        "logfile": "/var/log/pulseha.log"
+    }
+}
