+++ reset.sh
@@ -1,0 +1,4 @@
+sudo systemctl stop pulseha
+make
+sudo make install
+sudo systemctl restart pulseha
